backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
dj-database-url==3.1.0
djangorestframework-simplejwt
python-decouple==3.8
Pillow
psycopg2-binary
celery
redis
python-dateutil
django-cors-headers
django-filter
drf-spectacular
gunicorn